Remove Pause Feature
Remove pause to create a better one
