+++ Memorama Project/Assets/Scenes/CardsBehaviourTest.unity
@@ -375,19 +375,7 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 134128155}
   m_OnClick:
     m_PersistentCalls:
-      m_Calls:
-      - m_Target: {fileID: 781579862}
-        m_TargetAssemblyTypeName: ControlsManager, Assembly-CSharp
-        m_MethodName: ShowOrHidePauseMenu
-        m_Mode: 6
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
+      m_Calls: []
 --- !u!114 &134128155
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1484,6 +1472,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c08cac59c9860c7468d4fb5c6083bba6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  canvasGroup: {fileID: 0}
 --- !u!4 &553336878 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 7844202148939927265, guid: 8a2ed813e099f0e42be48aaadeb8f891, type: 3}
@@ -1992,15 +1981,13 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 826560799}
-  - component: {fileID: 826560800}
-  - component: {fileID: 826560801}
   m_Layer: 5
   m_Name: PauseMenuUI
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &826560799
 RectTransform:
   m_ObjectHideFlags: 0
@@ -2023,37 +2010,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!95 &826560800
-Animator:
-  serializedVersion: 3
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 826560798}
-  m_Enabled: 1
-  m_Avatar: {fileID: 0}
-  m_Controller: {fileID: 9100000, guid: cbb79b8c7c950114ba5b1a46f4c74c9a, type: 2}
-  m_CullingMode: 0
-  m_UpdateMode: 0
-  m_ApplyRootMotion: 0
-  m_LinearVelocityBlending: 0
-  m_WarningMessage: 
-  m_HasTransformHierarchy: 1
-  m_AllowConstantClipSamplingOptimization: 1
-  m_KeepAnimatorControllerStateOnDisable: 0
---- !u!114 &826560801
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 826560798}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 68f572366bba6904faaf480e23c07524, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &853856372
 GameObject:
   m_ObjectHideFlags: 0
@@ -2981,6 +2937,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c08cac59c9860c7468d4fb5c6083bba6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  canvasGroup: {fileID: 0}
 --- !u!1 &1184960827
 GameObject:
   m_ObjectHideFlags: 0
@@ -4886,8 +4843,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 560fdfc492df87b48aedf1ba782b0426, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  isPressing: 0
   value: 0
   holdDuration: 3
+  anim: {fileID: 0}
 --- !u!114 &2088547412
 MonoBehaviour:
   m_ObjectHideFlags: 0
