+++ Memorama Project/Assets/Scripts/Cards/CardBehaviour.cs
@@ -100,10 +100,9 @@ public class CardBehaviour : MonoBehaviour
 
     #region Private Methods
 
-    // Es llamado cuando el evento de "Pausa" es activado o desactivado. Esto evita que se puedan tocar las cartas cuando está el menú activado
-    void DisableCollider(bool boolValue)
+    void SetEnableCard(bool isEnable)
     {
-        boxCollider.enabled = boolValue;
+        boxCollider.
     }
 
     #endregion
@@ -112,12 +111,12 @@ public class CardBehaviour : MonoBehaviour
 
     private void OnEnable()
     {
-        ControlsManager.DisableActionsOnPauseMenu += DisableCollider;
+
     }
 
     private void OnDisable()
     {
-        ControlsManager.DisableActionsOnPauseMenu -= DisableCollider;
+
     }
 
     #endregion
