--- Memorama Project/Assets/Scripts/Others/ControlsManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 53928ce17f3dd3843999a47a975b8535
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
