--- Memorama Project/Assets/Scripts/Testing.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 560fdfc492df87b48aedf1ba782b0426
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
