--- Memorama Project/Assets/Scripts/Testing2.cs
@@ -1,37 +1,0 @@
-/*
-*********************************
- Script made by: Erick B. Gomez
- https://github.com/ErickBGomez
-*********************************
-*/
-
-using UnityEngine;
-
-public class Testing2 : MonoBehaviour
-{
-    #region Variables
-
-
-
-    #endregion
-
-    #region Unity Methods
-
-    void OnEnable()
-    {
-        Debug.Log("hola");
-    }
-
-    private void OnDisable()
-    {
-        Debug.Log("Adios");
-    }
-
-    #endregion
-
-    #region Public Methods
-
-
-
-    #endregion
-}
