--- Memorama Project/Assets/Scripts/Testing2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 52d70753b9f72ac4bbc5a68214688765
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
