--- Memorama Project/Assets/Scripts/UI/DisableCanvasGroupOnPause.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c08cac59c9860c7468d4fb5c6083bba6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
